requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2
pdfplumber==0.10.3
python-docx==0.8.11
streamlit==1.28.1
requests==2.31.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.1.3
Pillow==10.1.0
